+++ class-wicket-acc-main.php
@@ -9,7 +9,7 @@ namespace WicketAcc;
  * Plugin Name:       Wicket Account Centre
  * Plugin URI:        https://wicket.io
  * Description:       Customize WooCommerce my account features to build the Wicket Account Centre
- * Version:           1.1.3
+ * Version:           1.1.4
  * Author:            Wicket Inc.
  * Developed By:      Wicket Inc.
  * Author URI:        https://wicket.io
@@ -125,7 +125,6 @@ class WicketAcc
 			'includes/ray-stub.php',
 			'includes/helpers.php',
 			'includes/deprecated.php',
-			//'includes/routes/routes.php',
 		];
 
 		// Admin Classes